ok, wassup lib is not dependent on Miranda again
git-svn-id: http://svn.miranda-ng.org/main/trunk@5121 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/SmileyAdd/src/bkstring.h
@@ -159,7 +159,7 @@ public:
 	bkstring& operator += (const value_type _Ch)
 	{ return append(1, _Ch); }
 
-	value_type& operator[] (int ind) const
+	value_type& operator[] (size_t ind) const
 	{ return buf[ind]; }
 
 	friend bkstring operator+ (const bkstring& _Str1, const bkstring& _Str2)
